[LABS-488] Extract signing functionality to its own module
