+++ chia/wallet/wallet.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING, Any, ClassVar, cast
 
-from chia_rs import AugSchemeMPL, CoinSpend, G1Element, G2Element, PrivateKey
+from chia_rs import CoinSpend, G1Element, G2Element, PrivateKey
 from chia_rs.sized_bytes import bytes32
 from chia_rs.sized_ints import uint32, uint64, uint128
 from typing_extensions import Unpack
@@ -20,34 +20,16 @@ from chia.wallet.conditions import (
     CreateCoin,
     CreateCoinAnnouncement,
 )
-from chia.wallet.derivation_record import DerivationRecord
-from chia.wallet.derive_keys import (
-    MAX_POOL_WALLETS,
-    _derive_path,
-    _derive_path_unhardened,
-    master_sk_to_singleton_owner_sk,
-)
+from chia.wallet.derive_keys import _derive_path
 from chia.wallet.puzzles.p2_delegated_puzzle_or_hidden_puzzle import (
     DEFAULT_HIDDEN_PUZZLE_HASH,
-    calculate_synthetic_offset,
     calculate_synthetic_public_key,
     calculate_synthetic_secret_key,
     puzzle_for_pk,
     puzzle_hash_for_pk,
-    puzzle_hash_for_synthetic_public_key,
     solution_for_conditions,
 )
 from chia.wallet.puzzles.puzzle_utils import make_reserve_fee_condition
-from chia.wallet.signer_protocol import (
-    PathHint,
-    Signature,
-    SignedTransaction,
-    SigningInstructions,
-    SigningResponse,
-    Spend,
-    SumHint,
-    TransactionInfo,
-)
 from chia.wallet.transaction_record import TransactionRecord
 from chia.wallet.util.puzzle_decorator import PuzzleDecoratorManager
 from chia.wallet.util.transaction_type import CLAWBACK_INCOMING_TRANSACTION_TYPES
@@ -498,183 +480,3 @@ class Wallet:
 
     def hardened_pubkey_for_path(self, path: list[int]) -> G1Element:
         return _derive_path(self.wallet_state_manager.get_master_private_key(), path).get_g1()
-
-    async def sum_hint_for_pubkey(self, pk: bytes) -> SumHint | None:
-        pk_parsed: G1Element = G1Element.from_bytes(pk)
-        dr: DerivationRecord | None = await self.wallet_state_manager.puzzle_store.record_for_puzzle_hash(
-            puzzle_hash_for_synthetic_public_key(pk_parsed)
-        )
-        if dr is None:
-            return None
-        return SumHint(
-            [dr.pubkey.get_fingerprint().to_bytes(4, "big")],
-            calculate_synthetic_offset(dr.pubkey, DEFAULT_HIDDEN_PUZZLE_HASH).to_bytes(32, "big"),
-            pk,
-        )
-
-    async def path_hint_for_pubkey(self, pk: bytes) -> PathHint | None:
-        pk_parsed: G1Element = G1Element.from_bytes(pk)
-        index: uint32 | None = await self.wallet_state_manager.puzzle_store.index_for_pubkey(pk_parsed)
-        if index is None:
-            index = await self.wallet_state_manager.puzzle_store.index_for_puzzle_hash(
-                puzzle_hash_for_synthetic_public_key(pk_parsed)
-            )
-        root_pubkey: bytes = self.wallet_state_manager.root_pubkey.get_fingerprint().to_bytes(4, "big")
-        if index is None:
-            # Pool wallet may have a secret key here
-            if self.wallet_state_manager.private_key is not None:
-                for pool_wallet_index in range(MAX_POOL_WALLETS):
-                    try_owner_sk = master_sk_to_singleton_owner_sk(
-                        self.wallet_state_manager.private_key, uint32(pool_wallet_index)
-                    )
-                    if try_owner_sk.get_g1() == pk_parsed:
-                        return PathHint(
-                            root_pubkey,
-                            [uint64(12381), uint64(8444), uint64(5), uint64(pool_wallet_index)],
-                        )
-            return None
-        return PathHint(
-            root_pubkey,
-            [uint64(12381), uint64(8444), uint64(2), uint64(index)],
-        )
-
-    async def execute_signing_instructions(
-        self, signing_instructions: SigningInstructions, partial_allowed: bool = False
-    ) -> list[SigningResponse]:
-        root_pubkey: G1Element = self.wallet_state_manager.root_pubkey
-        pk_lookup: dict[int, G1Element] = (
-            {root_pubkey.get_fingerprint(): root_pubkey} if self.wallet_state_manager.private_key is not None else {}
-        )
-        sk_lookup: dict[int, PrivateKey] = (
-            {root_pubkey.get_fingerprint(): self.wallet_state_manager.get_master_private_key()}
-            if self.wallet_state_manager.private_key is not None
-            else {}
-        )
-        aggregate_responses_at_end: bool = True
-        responses: list[SigningResponse] = []
-
-        # TODO: expand path hints and sum hints recursively (a sum hint can give a new key to path hint)
-        # Next, expand our pubkey set with path hints
-        if self.wallet_state_manager.private_key is not None:
-            for path_hint in signing_instructions.key_hints.path_hints:
-                if int.from_bytes(path_hint.root_fingerprint, "big") != root_pubkey.get_fingerprint():
-                    if not partial_allowed:
-                        raise ValueError(f"No root pubkey for fingerprint {root_pubkey.get_fingerprint()}")
-                    else:
-                        continue
-                else:
-                    path = [int(step) for step in path_hint.path]
-                    derive_child_sk = _derive_path(self.wallet_state_manager.get_master_private_key(), path)
-                    derive_child_sk_unhardened = _derive_path_unhardened(
-                        self.wallet_state_manager.get_master_private_key(), path
-                    )
-                    derive_child_pk = derive_child_sk.get_g1()
-                    derive_child_pk_unhardened = derive_child_sk_unhardened.get_g1()
-                    pk_lookup[derive_child_pk.get_fingerprint()] = derive_child_pk
-                    pk_lookup[derive_child_pk_unhardened.get_fingerprint()] = derive_child_pk_unhardened
-                    sk_lookup[derive_child_pk.get_fingerprint()] = derive_child_sk
-                    sk_lookup[derive_child_pk_unhardened.get_fingerprint()] = derive_child_sk_unhardened
-
-        # Next, expand our pubkey set with sum hints
-        sum_hint_lookup: dict[int, list[int]] = {}
-        for sum_hint in signing_instructions.key_hints.sum_hints:
-            fingerprints_we_have: list[int] = []
-            for fingerprint in sum_hint.fingerprints:
-                fingerprint_as_int = int.from_bytes(fingerprint, "big")
-                if fingerprint_as_int not in pk_lookup:
-                    if not partial_allowed:
-                        raise ValueError(
-                            f"No pubkey found (or path hinted to) for fingerprint {int.from_bytes(fingerprint, 'big')}"
-                        )
-                    else:
-                        aggregate_responses_at_end = False
-                else:
-                    fingerprints_we_have.append(fingerprint_as_int)
-
-            # Add any synthetic offsets as keys we "have"
-            offset_sk = PrivateKey.from_bytes(sum_hint.synthetic_offset)
-            offset_pk = offset_sk.get_g1()
-            pk_lookup[offset_pk.get_fingerprint()] = offset_pk
-            sk_lookup[offset_pk.get_fingerprint()] = offset_sk
-            final_pubkey: G1Element = G1Element.from_bytes(sum_hint.final_pubkey)
-            final_fingerprint: int = final_pubkey.get_fingerprint()
-            pk_lookup[final_fingerprint] = final_pubkey
-            sum_hint_lookup[final_fingerprint] = [*fingerprints_we_have, offset_pk.get_fingerprint()]
-
-        for target in signing_instructions.targets:
-            pk_fingerprint: int = int.from_bytes(target.fingerprint, "big")
-            if pk_fingerprint not in sk_lookup and pk_fingerprint not in sum_hint_lookup:
-                if not partial_allowed:
-                    raise ValueError(f"Pubkey {pk_fingerprint} not found (or path/sum hinted to)")
-                else:
-                    aggregate_responses_at_end = False
-                    continue
-            elif pk_fingerprint in sk_lookup:
-                responses.append(
-                    SigningResponse(
-                        bytes(AugSchemeMPL.sign(sk_lookup[pk_fingerprint], target.message)),
-                        target.hook,
-                    )
-                )
-            else:  # Implicit if pk_fingerprint in sum_hint_lookup
-                signatures: list[G2Element] = []
-                for partial_fingerprint in sum_hint_lookup[pk_fingerprint]:
-                    signatures.append(
-                        AugSchemeMPL.sign(sk_lookup[partial_fingerprint], target.message, pk_lookup[pk_fingerprint])
-                    )
-                if partial_allowed:
-                    # In multisig scenarios, we return everything as a component signature
-                    for sig in signatures:
-                        responses.append(
-                            SigningResponse(
-                                bytes(sig),
-                                target.hook,
-                            )
-                        )
-                else:
-                    # In the scenario where we are the only signer, we can collapse many responses into one
-                    responses.append(
-                        SigningResponse(
-                            bytes(AugSchemeMPL.aggregate(signatures)),
-                            target.hook,
-                        )
-                    )
-
-        # If we have the full set of signing responses for the instructions, aggregate them as much as possible
-        if aggregate_responses_at_end:
-            new_responses: list[SigningResponse] = []
-            grouped_responses: dict[bytes32, list[SigningResponse]] = {}
-            for response in responses:
-                grouped_responses.setdefault(response.hook, [])
-                grouped_responses[response.hook].append(response)
-            for hook, group in grouped_responses.items():
-                new_responses.append(
-                    SigningResponse(
-                        bytes(AugSchemeMPL.aggregate([G2Element.from_bytes(res.signature) for res in group])),
-                        hook,
-                    )
-                )
-            responses = new_responses
-
-        return responses
-
-    async def apply_signatures(
-        self, spends: list[Spend], signing_responses: list[SigningResponse]
-    ) -> SignedTransaction:
-        signing_responses_set = set(signing_responses)
-        return SignedTransaction(
-            TransactionInfo(spends),
-            [
-                Signature(
-                    "bls_12381_aug_scheme",
-                    bytes(
-                        AugSchemeMPL.aggregate(
-                            [
-                                G2Element.from_bytes(signing_response.signature)
-                                for signing_response in signing_responses_set
-                            ]
-                        )
-                    ),
-                )
-            ],
-        )
