+++ chia/wallet/wallet_rpc_api.py
@@ -425,7 +425,7 @@ def tx_endpoint(
                 # deferring to parent action scope
                 return response
 
-            unsigned_txs = await self.service.wallet_state_manager.gather_signing_info_for_txs(
+            unsigned_txs = await self.service.wallet_state_manager.signer.gather_signing_info_for_txs(
                 action_scope.side_effects.transactions
             )
 
@@ -3790,7 +3790,7 @@ class WalletRpcApi:
         request: GatherSigningInfo,
     ) -> GatherSigningInfoResponse:
         return GatherSigningInfoResponse(
-            signing_instructions=await self.service.wallet_state_manager.gather_signing_info(request.spends)
+            signing_instructions=await self.service.wallet_state_manager.signer.gather_signing_info(request.spends)
         )
 
     @marshal
@@ -3800,7 +3800,9 @@ class WalletRpcApi:
     ) -> ApplySignaturesResponse:
         return ApplySignaturesResponse(
             signed_transactions=[
-                await self.service.wallet_state_manager.apply_signatures(request.spends, request.signing_responses)
+                await self.service.wallet_state_manager.signer.apply_signatures(
+                    request.spends, request.signing_responses
+                )
             ]
         )
 
@@ -3810,7 +3812,7 @@ class WalletRpcApi:
         request: SubmitTransactions,
     ) -> SubmitTransactionsResponse:
         return SubmitTransactionsResponse(
-            mempool_ids=await self.service.wallet_state_manager.submit_transactions(request.signed_transactions)
+            mempool_ids=await self.service.wallet_state_manager.signer.submit_transactions(request.signed_transactions)
         )
 
     @marshal
@@ -3819,7 +3821,7 @@ class WalletRpcApi:
         request: ExecuteSigningInstructions,
     ) -> ExecuteSigningInstructionsResponse:
         return ExecuteSigningInstructionsResponse(
-            signing_responses=await self.service.wallet_state_manager.execute_signing_instructions(
+            signing_responses=await self.service.wallet_state_manager.signer.execute_signing_instructions(
                 request.signing_instructions, request.partial_allowed
             )
         )
